@@ -2,25 +2,23 @@ package com.vsv.vova.androidreader;
 
 import android.content.Intent;
 import android.net.Uri;
-import android.support.v7.app.AppCompatActivity;
+import androidx.appcompat.app.AppCompatActivity;
 import android.os.Bundle;
-import android.support.v7.widget.LinearLayoutManager;
-import android.support.v7.widget.RecyclerView;
+import androidx.recyclerview.widget.LinearLayoutManager;
+import androidx.recyclerview.widget.RecyclerView;
 import android.view.View;
 import android.widget.Button;
-import com.vsv.vova.androidreader.Realm.ReaderRealm;
+
+import com.vsv.vova.androidreader.Room.Book;
 import com.vsv.vova.androidreader.adapter.OnItemClickListener;
 import com.vsv.vova.androidreader.adapter.RecyclerViewAdapter;
-import com.vsv.vova.androidreader.model.Book;
 
+import java.util.ArrayList;
 import java.util.List;
 
 import butterknife.BindView;
 import butterknife.ButterKnife;
 import butterknife.OnClick;
-import io.realm.RealmQuery;
-import io.realm.RealmResults;
-import io.realm.Sort;
 
 public class MainActivity extends AppCompatActivity implements OnItemClickListener {
 
@@ -32,7 +30,7 @@ public class MainActivity extends AppCompatActivity implements OnItemClickListen
     RecyclerView booksRecyclerView;
 
     private static Uri uri;
-    List<Book> bookList;
+    List<Book> bookList = new ArrayList<>();
 
     @Override
     protected void onCreate(Bundle savedInstanceState) {
@@ -45,13 +43,6 @@ public class MainActivity extends AppCompatActivity implements OnItemClickListen
     protected void onResume() {
         super.onResume();
         loadBookList();
-    }
-
-    private void openBook(Uri uri, long page){
-        Intent intent = new Intent(this, ReadActivity.class);
-        intent.putExtra("uri", uri.toString());
-        intent.putExtra("page", page);
-        startActivity(intent);
     }
 
     @OnClick(R.id.buttonFind)
@@ -72,25 +63,30 @@ public class MainActivity extends AppCompatActivity implements OnItemClickListen
         }
     }
 
+    private void openBook(Uri uri, int page){
+        Intent intent = new Intent(this, ReadActivity.class);
+        intent.putExtra("uri", uri.toString());
+        intent.putExtra("page", page);
+        startActivity(intent);
+    }
+
     private void loadBookList() {
+        ReaderRepository repo = new ReaderRepository(this.getApplication());
+        if(repo.getAll() != null){
+            bookList = repo.getAll();
+        }
+        recyclerInit(bookList);
+    }
 
-        RealmQuery<Book> bookRealmQuery = ReaderRealm.getRealm().where(Book.class);
-        bookList = bookRealmQuery.findAll();
-
-        bookList = ((RealmResults<Book>) bookList).sort("date" , Sort.DESCENDING);
-
-        LinearLayoutManager linearLayoutManager = new LinearLayoutManager(this);
-        booksRecyclerView.setLayoutManager(linearLayoutManager);
-
+    private void recyclerInit(List<Book> bookList){
+        booksRecyclerView.setLayoutManager(new LinearLayoutManager(this));
         RecyclerViewAdapter recyclerViewAdapter = new RecyclerViewAdapter(bookList);
         recyclerViewAdapter.setOnItemClickListener(this);
-
         booksRecyclerView.setAdapter(recyclerViewAdapter);
-
     }
 
     @Override
     public void onItemClick(View view, int position) {
-       openBook(bookList.get(position).getUri() ,bookList.get(position).getPage());
+       openBook(Uri.parse(bookList.get(position).getUri()) ,bookList.get(position).getPage());
     }
 }
