@@ -1,15 +1,15 @@
 package com.vsv.vova.androidreader.adapter;
 
 import android.annotation.SuppressLint;
-import android.support.annotation.NonNull;
-import android.support.v7.widget.RecyclerView;
+import androidx.annotation.NonNull;
+import androidx.recyclerview.widget.RecyclerView;
 import android.view.LayoutInflater;
 import android.view.View;
 import android.view.ViewGroup;
 import android.widget.TextView;
 
 import com.vsv.vova.androidreader.R;
-import com.vsv.vova.androidreader.model.Book;
+import com.vsv.vova.androidreader.Room.Book;
 
 import java.util.List;
 
